+++ src/graphQLPaginationType.ts
@@ -1,0 +1,9 @@
+import { GraphQLInputObjectType, GraphQLInt } from "graphql";
+
+export default new GraphQLInputObjectType({
+    name: "PaginationType",
+    fields: {
+        limit: { type: GraphQLInt },
+        skip: { type: GraphQLInt }
+    }
+});